CORE-2686 Refactor updateApplication success handling.
Refactored the EditCompleteCallback in CatalogMainAdminPanel so that it extends the AdminServiceCallback, and the updateApp method so that it accepts the app JSON object directly.
Updated EditAppDetailsPanel so that it doesn't close automatically on update failures.
Added an addCancelButtonSelectionListener method to EditAppDetailsPanel that the enclosing dialog panel in CatalogMainAdminPanel can use to close itself when the Cancel button is clicked.