Added idrop close prompt to Simple Download button.
Refactored the IDropLiteAppletWindow MessageBox prompt into a promptRemoveApplet method that will execute a given GWT Command when the user selects "Yes".
Added an abstract SimpleFormSelectionListener for the Simple Download and Upload form buttons that will implement their own removeAppletConfirmed actions, called when the user confirms the close-applet-prompt.
Re-worded the idropLiteCloseConfirmMessage text to refer only to the applet instead of the upload/download window.