CORE-2534 Updates for handling AnalysisSelectEvent.
Added an AnalysisSelectEventHandler (for AnalysisSelectEvents sent by selecting App search results) in DECatalogWindow that calls a new DECatalogWindow::launchDECatalogWindow static method, which will message the Window Manager to show (or open) the DECatalogWindow, with a CatalogWindowConfig that tells the window to select the category and app set in the config.
Added an overridden cleanup method to DECatalogWindow that will remove its event handlers.
Added a categoryID field to CatalogWindowConfig.
Updated WindowConfigFactory to build a CatalogWindowConfig if the config's type is the deCatalog string constant (as set in launchDECatalogWindow).