CORE-2534 Updated launching DECatalogWindow in DefaultWorkspaceView.
Updated DefaultWorkspaceView to use the DECatalogWindow::launchDECatalogWindow static method to open this window and select an App, instead of creating a new window instance and showing it.
This new static method will message the Window Manager with an appropriate window config, so that only one instance of this window is created, and it will have a toolbar button for restoring it when it's minimized.