CORE-2534 Updates for App Catalog remote search.
Added a belphegor_catalog_panel tag for AnalysisSelectEvents.
Added a search-apps URL property and accessor in ToolIntegrationAdminProperties.
Added a stub for searchAnalysis in AppTemplateAdminServiceFacade.
Updated CatalogAdminPanel for the new CatalogMainAdminPanel constructor, and added an AnalysisSelectEvent handler.
Updated CatalogMainAdminPanel for the new BaseCatalogMainPanel constructor, and added a getTemplateService method to access its protected AppTemplateAdminServiceFacade.
Added CSS styles for the App Catalog search results.