CORE-2886 Fixed simple file uploads in secured environments.
Updated FileUploadServlet so that the DataApiServiceDispatcher is used for file upload requests as well as URL import requests.
Since the Scruffian service is running under Jetty, the large SAML headers sent by the DEServiceDispatcher will cause Jetty to return a 413 HTTP error.
Removed the "user" part from the MultiPartServiceWrapper since the DataApiServiceDispatcher will automatically add the user to the request.