CORE-2619 Added file preview and manifest error handling.
Refactored the DiskResourceServiceCallback::onSuccess method and related EventJSONFactory and MessageDispatcher processing code into a new DiskResourceActionCallback class.
Updated the getManifest and getData methods in FileEditorServiceFacade to use a DiskResourceServiceCallback, instead of an AsyncCallback, so that error messages can be automatically parsed and displayed.
Added an unableToRetrieveFileData default error message for file preview and raw data request errors.