CORE-2932 Fixed move exceptions in MyDataWindow.
Updated the folderMove and fileMove methods in MyDataWindow to delete the moved disk resources from the main data panel all at once with the deleteDiskResources method.
Removing these resources one at a time caused the selection model to fire a changed event each time, then when there was a single item left to remove, either a directory or file-manifest endpoint request would be triggerred for it; but since that item was already moved, an exception would be thrown for the request.