CORE-2993 Allow 2 idrop windows to be open at once.
Updated the static launch-window methods in IDropLiteAppletWindow to use different tags for windows in upload and download mode.
Updated WindowFactory to accomodate the new tags.
This will allow an upload applet window to be open at the same time as a download applet window.
Added a notice to the applet window in download mode to let the user know that only 1 window in download mode can be open at a time.