CORE-2514 Data nav tree lazy-loading fix.
Refactored the DataNavigationPanel::loadFolderAndSelect method added in commit 95ce90e188e3df32abc9c55eeb0344140dd8584e, renaming it to reloadFolderThenExpand.
The previous loadFolderAndSelect method assumed the next child to load was the final folder to select, but there may be deeper children along the path still to load.
Now this method just reloads the given folder, removing then readding it to the tree store, then reattempts to expand it in order to continue the lazy-loading process.
This ensures it's in sync with the backend and that the GXT tree model's state is not stale.
Also added an onFailure method to this service callback that will remove the lazy-loading callback from the tree-loader before displaying its error message.
Reverted the mistaken change made to MyDataWindow::show in commit 95ce90e188e3df32abc9c55eeb0344140dd8584e.