CORE-3012 Added TitoWindow launched by WindowManager.
Refactored the tito window code, that was built and displayed in the CatalogMainPanel::buildNewAnalysis menu action, into a TitoWindow class.
Added a launch method to this class that will dispatch the appropriate message so that the WindowManager can build and display this window via the WindowFactory.
Added a createApps display string for this window's title.