CORE-3012 Added TitoController and TitoWindowConfig.
TitoController now listens to the events that the ApplicationController in the tito repo used to listen to, then launches a TitoWindow with a TitoWindowConfig configured by the event listener.
Added an instance of TitoController to DesktopView.
Added a new launch method to TitoWindow that will dispatch the approiate window launch message with JSON for a TitoWindowConfig.
Added a cleanup method to TitoWindow that will call the tito ApplicationLayout cleanup method.
TitoWindow will now call the appropriate tito ApplicationLayout method according to its TitoWindowConfig.