CORE-210 Added WindowDispatcher to launch windows.
Refactored the buildConfigPayload method in WindowConfigFactory to a buildWindowConfig method that only returns JSON for the "config" key in window payload JSON objects, since this object is now built by the WindowDispatcher::dispatchAction method.
Refactored the static launch methods in DECatalogWindow and IDropLiteAppletWindow to use the new WindowConfigFactory::buildWindowConfig method and WindowDispatcher.