CORE-3029 Fix for extra Nibblonian error codes.
If a Nibblonian error code is not defined in the UI's ErrorCode enum, then trying to parse that code will throw an exception, which was causing no error dialog to display to the user at all.
Updated DiskResourceServiceCallback::onFailure to catch these enum parse exceptions and to just display the default error message to the user and to include the string value of the error code in the error dialog details.