CORE-3057 idrop-lite applet workaround for Chrome and IE9.
Updated IDropLiteAppletWindow to auto-minimize in Chrome and IE9 when the window loses focus.
This is a workaround for the fact that the applet remains on top of all other content in these browsers, even if opening another GXT window or dragging one over this window.
The GXT forums suggestion of setting GXT.useShims to true and setting the idrop-lite window and applet content panel setShim to true only worked in these browsers, but the applet still covered dialogs and tooltips (auto-previews in the data window).