CORE-3119 Added Simple Download link to DataActionsPanel.
Refactored the simple download links from the idrop-lite window into their own SimpleDownloadWindow.
Added a "Simple Download" link to the DataActionsPanel that will open this window, in addition to the "Simple Download" button in the idrop-lite window.
If only 1 file is selected in the Data window, then the download will begin immediately.
Refactored the file download logic into a simpleDownload method in FolderServiceFacade.