CORE-3224 Updated properties file in SimpleMessageSender.
SimpleMessageSender was still trying to lookup email settings from a tito.properties file.
Updated SimpleMessageSender to use discoveryenvironment.properties instead.
Additionally, the exception from the SimpleMessageSender constructor was not getting caught by the NewToolRequestServlet, so the SimpleMessageSender instantiation has now been moved inside the try-catch block.
The SimpleMessageSender was throwing a java.lang.NullPointerException by trying to open a missing file, and this type of exception returns null for its getMessage method, which would cause the servlet to return an empty JSON message that the UI assumed was a success.
Added an exception message null-check in the json error response, so that at least the name of the exception is returned.