CORE-2597 Added deployed components listing dialog.
Removed the AppDCDetailsPanel from the East panel of the DECatalogWindow.
Refactored AppDCDetailsPanel into an IPlantDialogPanel and added a Loading mask around the TemplateServiceFacade::getDCDetails call.
Added a "View Integrated Tools" button to CatalogMainPanel that launches an IPlantDialog with the AppDCDetailsPanel for the selected App.
Updated TemplateServiceFacade::getDCDetails to use the actual service URL.