CORE-3442 Added WindowFocusDropTarget for windows and taskbar buttons.
Added a WindowFocusDropTarget class that listens to drag events over some Component, in order to focus (bring to front) some window after a small delay, when the user drags then hovers an item over that target Component.
Added private classes to IPlantTaskButton and IPlantWindow that extend WindowFocusDropTarget in order to focus their respective windows after a drag-and-hover delay.