CORE-3505 Updated WizardWindows to support many groups.
Updated WizardWindow so that it's maximizable and uses a border layout for its contents.
Updated WizardBuilder so that the ContentPanel it builds supports scrolling, to accommodate many property groups that may not fit in the default window height.
Also removed unnecessary configs (such as setting AutoHeight to false) in WizardWindow, WizardBuilder, and AccordionChildPanel.